get it working with gcc 3.2 and recent binutils

+++ kmalloc.h
@@ -3,5 +3,5 @@ void *krealloc(void *_r, ulong n);
 void kfree(void *_r);
 void __heap_trace(void);
 
-void *memcpy(void *dst, const void *src, ulong len);
-void *memset(void *dst, int val, ulong len);
+void *memcpy(void *dst, const void *src, size_t len);
+void *memset(void *dst, int val, size_t len);
